Revert "Revert "Rename package for parsing items.""
This reverts commit 00296e53bac2566c634fa1ab8c4318b9663dac83.
