+++ src/main/java/org/rumbledb/itemparsing/RowToItemMapper.java
@@ -18,40 +18,24 @@
  *
  */
 
-package sparksoniq.io.json;
+package org.rumbledb.itemparsing;
 
-import org.apache.spark.api.java.function.FlatMapFunction;
+import org.apache.spark.api.java.function.Function;
+import org.apache.spark.sql.Row;
 import org.rumbledb.api.Item;
-import sparksoniq.jsoniq.item.ItemFactory;
+import sparksoniq.jsoniq.runtime.metadata.IteratorMetadata;
 
-import java.util.Iterator;
-
-public class StringMapper implements FlatMapFunction<Iterator<String>, Item> {
-
+public class RowToItemMapper implements Function<Row, Item> {
 
     private static final long serialVersionUID = 1L;
+    private final IteratorMetadata metadata;
 
-    public StringMapper() {
+    public RowToItemMapper(IteratorMetadata metadata) {
+        this.metadata = metadata;
     }
 
     @Override
-    public Iterator<Item> call(Iterator<String> stringIterator) throws Exception {
-        return new Iterator<Item>() {
-            @Override
-            public boolean hasNext() {
-                return stringIterator.hasNext();
-            }
-
-            @Override
-            public Item next() {
-                Item stringItem = ItemFactory.getInstance().createStringItem(stringIterator.next());
-                return stringItem;
-            }
-
-            @Override
-            public void remove() {
-                throw new UnsupportedOperationException();
-            }
-        };
+    public Item call(Row row) throws Exception {
+        return ItemParser.getItemFromRow(row, metadata);
     }
 }
