+++ src/main/java/org/rumbledb/itemparsing/StringToStringItemMapper.java
@@ -18,24 +18,40 @@
  *
  */
 
-package sparksoniq.io.json;
+package org.rumbledb.itemparsing;
 
-import org.apache.spark.api.java.function.Function;
-import org.apache.spark.sql.Row;
+import org.apache.spark.api.java.function.FlatMapFunction;
 import org.rumbledb.api.Item;
-import sparksoniq.jsoniq.runtime.metadata.IteratorMetadata;
+import sparksoniq.jsoniq.item.ItemFactory;
 
-public class RowToItemMapper implements Function<Row, Item> {
+import java.util.Iterator;
+
+public class StringToStringItemMapper implements FlatMapFunction<Iterator<String>, Item> {
+
 
     private static final long serialVersionUID = 1L;
-    private final IteratorMetadata metadata;
 
-    public RowToItemMapper(IteratorMetadata metadata) {
-        this.metadata = metadata;
+    public StringToStringItemMapper() {
     }
 
     @Override
-    public Item call(Row row) throws Exception {
-        return JiqsItemParser.getItemFromRow(row, metadata);
+    public Iterator<Item> call(Iterator<String> stringIterator) throws Exception {
+        return new Iterator<Item>() {
+            @Override
+            public boolean hasNext() {
+                return stringIterator.hasNext();
+            }
+
+            @Override
+            public Item next() {
+                Item stringItem = ItemFactory.getInstance().createStringItem(stringIterator.next());
+                return stringItem;
+            }
+
+            @Override
+            public void remove() {
+                throw new UnsupportedOperationException();
+            }
+        };
     }
 }
