+++ src/main/java/sparksoniq/jsoniq/runtime/iterator/HybridRuntimeIterator.java
@@ -24,9 +24,10 @@ import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 import org.rumbledb.api.Item;
+import org.rumbledb.itemparsing.ItemParser;
+import org.rumbledb.itemparsing.RowToItemMapper;
+
 import sparksoniq.exceptions.IteratorFlowException;
-import sparksoniq.io.json.JiqsItemParser;
-import sparksoniq.io.json.RowToItemMapper;
 import sparksoniq.jsoniq.ExecutionMode;
 import sparksoniq.jsoniq.runtime.metadata.IteratorMetadata;
 import sparksoniq.semantics.DynamicContext;
@@ -37,7 +38,7 @@ import java.util.List;
 public abstract class HybridRuntimeIterator extends RuntimeIterator {
 
     private static final long serialVersionUID = 1L;
-    protected JiqsItemParser parser;
+    protected ItemParser parser;
     protected List<Item> result = null;
     private int currentResultIndex = 0;
 
@@ -48,7 +49,7 @@ public abstract class HybridRuntimeIterator extends RuntimeIterator {
     ) {
         super(children, executionMode, iteratorMetadata);
         fallbackToRDDIfDFNotImplemented(executionMode);
-        this.parser = new JiqsItemParser();
+        this.parser = new ItemParser();
     }
 
     protected boolean implementsDataFrames() {
