+++ src/main/java/sparksoniq/jsoniq/tuple/FlworTuple.java
@@ -28,8 +28,9 @@ import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 import org.rumbledb.api.Item;
+import org.rumbledb.itemparsing.RowToItemMapper;
+
 import sparksoniq.exceptions.OurBadException;
-import sparksoniq.io.json.RowToItemMapper;
 import sparksoniq.jsoniq.runtime.metadata.IteratorMetadata;
 import sparksoniq.spark.SparkSessionManager;
